Refactor player and enemy logic into components
